Extract utility fns from integration-tests/tests.ts
